Add an outer reduction test for TMA circular buffering (#3516)
Inner reduction uses block reduction, which implies block sync, which is
not supported in warp specialization. So adding another test so that we
can test warp specialization with reduction. Note that we will not
support persistent + warp specialization. I can not think of a way to
writing persistent kernel without using block syncs. We can of course
replace that block sync with mbarrier, which provides finer-grained
control to only synchronizing on the compute warp, but that needs labor
and is not what we are prioritizing now.