Skip loop promotion analysis a little more aggressively (#3526)
Currently we check all loop groups, no matter if they correspond to loop
IDs. For loop promotion, we only need to consider loop domains when
checking if the full loop promotion analysis can be skipped. This
refinement was necessary when playing with scheduling approaches for
resize, although it is not currently necessary in #3425. I think this is
generally better than the current method.