Fix domain validation (#3514)
Fixes #3502 

To fix #3502, I revisited `compareDomains` and decided to start from
scratch. `compareDomainWithReference` is the new function to verify the
validity of a domain.

As the name suggests, this new function assumes one of the input domains
is known to be a domain consisting of independent non-overlapped iter
domains. That makes the analysis much simpler and less expensive. The
assumption should be fine as we can always use the logical domain as a
reference.

I initially attempted to just fix `compareDomains` with no particular
assumption like having a reference
(https://github.com/NVIDIA/Fuser/pull/3514/commits/46089ef306a1b6da6e7d63d1962dedc4679d81e3),
but it turned out to be a very expensive analysis, so I scrapped it.