add scheduler_tools::scheduleLoopDomainsBy (#3523)
Extracted from #3425 

A simple scheduling utility that replays a single transform expr on the
loop domain of a given tensor. It is conceptually similar to the
existing transform propagator, but this interface can also replay a
transform expr not just as a forward expr but also as a backward expr.

The existing `scheduler_tools::scheduleLoopDomainsLike` can also be used
to propagate loop domains across tensors, but I encountered with a
couple of issues due to the resize mapping [issue](#3455). This
`scheduleLoopDomainsBy` gives more explicit control to set loop domains
with resize ops.