Resize war follow-up fix (#3530)
This is a second attempt to fix #3505. The first attempt is #3515. As
mentioned
[here](https://github.com/NVIDIA/Fuser/issues/3505#issuecomment-2519021709),
the first fix isn't sufficient when an expr has multiple resized inputs,
like concat. The actual condition we need to check is between each
producer and consumer pair, not between producers, so this second
attempt is just changing how we check the condition.