Update all reduction primitives to take a block_dim parameter (#3536)
Stacked on https://github.com/NVIDIA/Fuser/pull/3541

This PR is mostly mechanical. It just update all the reduction
primitives with a new parameter `block_dim`, which is just `blockDim`
for a kernel without warp specialization. For a kernel with warp
specialization, we use a value for `block_dim`, which is the block
dimension of the compute warps.

For better performance, it is beneficial if we can know whether we are
using the `blockDim` or a custom value at compile time. To achieve this,
I added a new helper class:

```C++
struct DefaultBlockDim {
  const uint32_t x, y, z;
  __device__ DefaultBlockDim() : x(blockDim.x), y(blockDim.y), z(blockDim.z) {}
  __device__ operator dim3() const {
    return blockDim;
  }
};
```
When a function takes `block_dim`, its type will be a template type
`BlockDimT`, which is either `DefaultBlockDim` or `dim3`, where
`DefaultBlockDim` means we are using `blockDim`, and `dim3` means we are
using the provided custom value.

With this PR, I would consider all our reductions and broadcasts
compatible with warp specialization. I expect them to just work with
warp specialization. In the future, we may want to use TMA + warp
specialization for reduction and persistent kernels.