WAR for supporting the rotation + residual pattern (#3555)
Stacked on top of #3549. This is also a WAR for #3455 and necessary to
schedule RoPE-like rotation patterns.

Because of the issue, a tensor may have two IDs that are exactly mapped.
For example, when an ID is sliced to half and then is padded back to the
same size, and the final output ID is used with the initial input ID,
the initial input and the final output IDs get mapped together. This can
make it difficult to use `scheduleLoopDomainsLike`. For example, if a
reference has a split that is done with the final output ID, and we want
to replay the split on other tensors, it becomes ambiguous whether the
split is done with the initial input or the final output since both are
exactly mapped.

To avoid this ambiguity, this PR adds a flag to indicate that we just
want to update the current loop domain with a reference domain. As seen
in the added tests, this flag is used to propagate the scheduling of a
reference tensor once all resize ops are propagated to inputs.
Specifically, the overall scheduling follows this pattern:

1. Propagate all slice and pad ops to fusion inputs
2. Pick and schedule a reference tensor
3. Propagate the scheduling of the reference tensor to the other tensors

`scheduleLoopDomainsLike` with the flag is used at step 3. For that
step, we know that we don't need to schedule each tensor with a complex
replay path, like some backward ops followed by some other forward ops,
but we just need to update the current loop domain by replaying the diff
with the reference domain.