Allow inlining past loop broadcasts for MmaOp (#3416)
Stacked on #3414 

This PR enables us to inline an MmaOp properly when its inputs are
missing broadcast dimensions. We do this by always allowing inlining
past loop broadcasts or their transforms. For example
```
tv0:
  logical [ iS1{i0} ]
  loop [ iS1{i0} bS5{1} ]
tv1:
  logical [ iS2{i1} ]
  loop [ bS6{1} iS2{i1} ]
tv2 = foo(tv0, tv1)
  logical [ iS3{i0} iS4{i1} ]
```
As long as the operation `foo` properly maps its arguments despite the
missing logical dimensions (as `MmaOp` does as of #3391), then we should
be able to fully inline this case because the loop broadcasts `bS5` and
`bS6` are imaginary in the sense that they don't impact indexing.