Add cluster_dims param to MatmulParams
Following #3557 we can specify the cluster size for our fusions.
Currently we don't do anything explicitly with CGAs, but this can help
guarantee that tiles are scheduled onto GPCs in pairs. Each GPC has a
number of TPCs, each of which holds 2 SMs, so this lets us take
advantage of caching at the TPC and GPC level for operand loads, in
addition to L2.