Add MatmulParams::cluster_dims parameter (#3574)
Following #3557 we can specify the cluster size for our fusions.
Currently we don't do anything explicitly with CGAs, but this can help
guarantee that tiles are scheduled onto GPCs in pairs. Each GPC has a
number of TPCs, each of which holds 2 SMs, so this lets us take
advantage of caching at the TPC and GPC level for operand loads, in
addition to L2.

This PR enables this with a default size of `{2, 1, 1}` for the Hopper
scheduler. The parameter is ignored in the Ampere scheduler.

It is not yet plumbed into the heuristic plugin API yet. I thought maybe
we should wait until we have more parameters related to CGAs to do that.