Fix double-scheduling of dc in HSS hopper matmuls (#3590)
When we do not have an epilogue (not even a cast), it might be the case
that the original `MmaOp` has output which is a Fusion output. In this
case the cached output which we often call `dc` is actually an
`mma_result`. Currently this causes us to schedule that tensor once in
`scheduleMmaResults` then again in `scheduleEpilogue`, leading to an
esoteric error (see included test). This PR simply skips scheduling
those tensors directly if they are already known to be mma results.