Enable the Loop option of IdModle for resize (#3588)
Followup to #3567. I just found the Loop option is also necessary. With
this option, the inlining analysis uses IdModel to understand if
inlining is possible. The loop generation lowering pass also uses
IdModel loop promotion to figure out which iter domains to use for each
`ForLoop` node. The latter is not necessary for the resize scheduler at
this moment, though.