Enable translation of Hopper matmuls (#3440)
This enables Hopper matmul in our automatic scheduler by translating
them without introducing new broadcasts. Specifically:

1. Update `mma_utils::MatmulPattern::translateToMmaOp` to optionally
avoid intermediates by using an `MmaOp::AxisMapping`. Enable this option
when the target arch is not Ampere or Turing.
3. Unguard some tests in `test_translate_mma.cpp`

This does not update the default heuristic or change the `canSchedule`
checks. See #3579 for that follow-up PR

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>
Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: Jingyue Wu <wujingyue@gmail.com>
Co-authored-by: nsarka <nsarkauskas@nvidia.com>
Co-authored-by: Protonu <pbasu@nvidia.com>
Co-authored-by: samnordmann <snordmann@nvidia.com>