Initial resize scheduler (#3556)
This is a very preliminary version of a new scheduler mainly targeted
for RoPE. I will incrementally extend this scheduler to be more flexible
and performant, but for now it only handles a fusion that has pointwise
ops and a single `Resize`-based tensor op such as `SliceOp` and `PadOp`.
The scheduling strategy is currently pretty naive too and is manually
demonstrated at #3549 and #3555, but the main point is that inputs of
resize-based tensor ops like `SliceOp` or `PadOp` no longer need to have
their inputs as fusion inputs.

The new scheduler is currently placed after the reduction scheduler and
before the transpose and pointwise schedulers:

```
SchedulerType::ExprEval,
    SchedulerType::NoOp,
    SchedulerType::Matmul,
    SchedulerType::Reduction,
    SchedulerType::Resize, <-- New
    SchedulerType::Transpose,
    SchedulerType::PointWise,
    SchedulerType::InnerPersistent,
    SchedulerType::OuterPersistent,
    SchedulerType::InnerOuterPersistent};
```

https://github.com/NVIDIA/Fuser/pull/3556/files#diff-c0d261d44c61935fa2d5398f0ac52bd6ea077c6892fb5629c03a425a55fc32f2R64-R74

There are several small changes with some of the existing tests, mainly
those on segmentation and alias support since this new scheduler may
change how a fusion is segmented when resize is used. There's one thing
I haven't addressed
(https://github.com/NVIDIA/Fuser/pull/3556#discussion_r1879300006),
which I'm tracking with a separate issue.