Translate MatmulPattern to MmaOp using OptInDispatch (#3593)
This PR should not introduce any behavior change. I am refactoring
`MatmulPattern::translateToMmaOp` to use `OptInDispatch` to handle the
different types of pattern like mul+sum, `LinearOp`, and `MatmulOp`.

`OptInDispatch` differs from the more common `OptOutDispatch` by
throwing an error when we encounter unhandled statements. This is
appropriate in this case because we have a fixed set of supported
patterns and we do not wish to fall through to some default translation
in the case that we encounter an unsupported pattern; it is better to
throw an error.

Original suggestion to use separate methods for each pattern:
https://github.com/NVIDIA/Fuser/pull/3440#pullrequestreview-2503462655.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>
Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: Jingyue Wu <wujingyue@gmail.com>
Co-authored-by: nsarka <nsarkauskas@nvidia.com>
Co-authored-by: Protonu <pbasu@nvidia.com>
Co-authored-by: samnordmann <snordmann@nvidia.com>