pointwise scheduler fails to validate reference tv (#3513)
Fixes: #3512 

When picking reference tv, pointwise scheduler fails to validate that
the transformation on reference tv can be safely propagated to all
outputs in the fusion. The issue occurs when an IterDomain that's not in
the reference tv is merged with another dimension in the output tv,
preventing the merge on reference tv to be propagated to the target.

This PR adds an optional check `areAllOutputIdsMappedTo` in
`nvfuser::pointwise_utils::DomainMap::isValidReference`

The added check in this PR checks that all source producer IterDomain
producing the IterDomain on outputs are covered by reference tv. This is
safe for pointwise scheduler, since the scheduler checks that there's no
reversible view present in the fusion.

The check is optional and is disabled by transpose scheduler, where the
reference_tv is not supposed to cover the entire fusion, but rather a
subset of fusion IO tensors. We should extent that in future PRs.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: Jacob Hinkle <1454944+jacobhinkle@users.noreply.github.com>