Test horizontal matmul fusion in Llama2FFN test (#3610)
This removes some barriers to horizontal fusion and updates the test
which is currently Ampere-only.

Note that most of the horizontal fusion code hasn't been exercised much
so we might continue hitting small snags as we start using it more. My
intention with this PR is to test it automatically by modifying the
test. Likewise, we will need changes to the canSchedule checks and
default heuristics to ensure sane behavior when doing horizontal
fusions, so there will likely be more PRs of this flavor soon.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>