Allow fusion of multiple exclusive resize ops (#3611)
Followup to #3556. Currently, the resize scheduler is only allowed with
a single slice or pad. This PR allows for fusing multiple ops as long as
they don't conflict. Please see the
[comment](https://github.com/NVIDIA/Fuser/pull/3611/files#diff-b066c49d399243d3be36a44f1221490b9a2f50e41074feab836bc9bb6ee71180R25-R100)
for `getNonExclusiveResizeInfo`.

In this PR, if there's a conflict, the fusion is simply rejected. A
followup PR will address this limitation by replicating computations.