Host IR: make stream synchronization non blocking (#3608)
# What

Make stream synchronization non-blocking from the CPU point of view

# Why

Needed for achieving overlap in 
- https://github.com/NVIDIA/Fuser/pull/3606

before this patch:
![Screenshot 2024-12-18 at 12 08
25](https://github.com/user-attachments/assets/f5c84282-ea85-4cb8-8a60-538cd91cfa1c)
after this patch
![Screenshot 2024-12-18 at 12 08
05](https://github.com/user-attachments/assets/25537a5d-3e33-4ff8-baf4-4f013c1ed230)


# How 

Before this patch, the host IR `Synchronize` would call
`c10::synchronize()` on the cuda stream, which makes the CPU blocks
until stream completion. With this patch, we synchronize the current
stream with a given stream through a `cudaEvent` and the API
`cudaStreamWaitEvent`.