Allow matmul heuristic plugin to set cluster dimensions (#3634)
This allows heuristic plugins to set cluster dimensions. By default, the
cluster dims are set to {1, 1, 1}, which disables this feature, so if a
plugin does not explicitly handle cluster dims, then it will just not
make use of this feature. This is necessary because setting to an
invalid value can cause a launch failure. This PR also prints the
cluster dims in `MatmulParams::toString`.