Resolve conflicts by recomputation (#3625)
Stacked on top of #3611 

This PR resolves the conflicts found by the analysis added at #3611 by
recomputing slice/pad input tensors. With this, fusions like
`ResizeSchedulerTest.SliceRotateCatResidual` can be scheduled as a
single kernel by the resize scheduler.

Recomputation is not the only possible way to resolve conflicts. We
could do, for example, cache a block of a input tensor such that
multiple uses of the input can be done by just using the block. This
would look more like a produce-based scheduling approach. I prototyped
that approach [here](#2869), but it didn't perform well for RoPE.

---------

Co-authored-by: Jacob Hinkle <1454944+jacobhinkle@users.noreply.github.com>