Create SetMaxNReg and Return kir nodes (#3667)
This PR adds `kir::SetMaxNReg` and `kir::Return` nodes for register
sharing with warp specialization.

`kir::SetMaxNReg` corresponds with
[setmaxnreg](https://docs.nvidia.com/cuda/parallel-thread-execution/#miscellaneous-instructions-setmaxnreg)
ptx instruction to add a compiler hint to increase or decrease the
number of registers in a warp group.

`kir::Return` corresponds with an empty `return` statement. It is
necessary to terminate the load warps early. Otherwise, cuda kernel will
have poor performance.