C++ layer to implement Triu (#3631)
This is a C++ API to implement `triu`. The PR
(https://github.com/NVIDIA/Fuser/pull/3637) stacked on top of this used
it to create a Python interface.

Another way of using this may be to use the components of `triu` such as
`iota`, `broadcast`, `le` and `where` from Thunder directly bypassing
the need for a C++ implementation. As future work this commit/PR may be
removed in favor of a Thunder only implementation.