Don't try to index broadcast IDs as they should always be zero and there may not be a valid path from loop IDs (#3690)
Fixes #3688 

The error of #3688 is because some of the broadcast logical IDs are not
reachable when indexing. Specifically, this happened when indexing the
logical IDs of a consumer tensor of a pad op.
https://github.com/NVIDIA/Fuser/blob/main/csrc/device_lower/pass/index.cpp#L2698-L2699

When a consumer has broadcast logical IDs, they may not be reachable
from the loop domain when inlined, and that's why the indexing traversal
fails as seen in #3688. Note that for usual indexing paths such as
`getLinearIndex` and `getPredicates`, they always filter out broadcast
IDs from indexing because they should always use zero.

This PR always excludes broadcast IDs from indexing by changing
`computeIndex`, which is the common indexing interface for all use cases
including `getLinearIndex`, `getPredicates`, `getIndexFor`, etc.

To be honest, I'm not 100% confident if we should really always ignore
broadcast IDs, but I'm not aware of any concrete counter cases. Given
that, this seems to be the most reasonable fix to me.