Matmul with DID loop split (#3651)
This PR modifies the `hasTrivialAllocationDomain` to consider if the
tensorview has a DID loop split. In this case, we compare the
corresponding iterdomains for logical and allocation domain across all
but the sharded logical axis.

Note: This does not guarantee that `MatmulOp` with non-trivial stride
order will work for DID loop split. I suspect it will require some
additional changes to the `MatmulOp::evaluate` method.