expand removing consecutive cast to handle meta operations in between (#3644)
Existing ConsecutiveCast optimization pass only optimize a consecutive
cast operations. This PR expand ConsecutiveCast pass to handle cases
where a chain of cast operations is broken by a meta operation in the
middle.

e.g.
```
T1 = castOp(T0, fp32)
T2 = squeeze(T1)
T3 = castOp(T2, fp16)
```

The existing pass wouldn't be able to cancel out the two casts, because
they are separated by the squeeze operation.

In this PR, before we trace back from the last CastOp for the chain of
casts, we look at the input to the cast operation. If it's a *movable*
meta operation, we swap the order of the meta op and the cast op first,
then we resume the chain look up on consecutive casts.