Add extra validation for IdModel (#3746)
As suggested by @jjsjann123, this PR adds an extra error check for the
Exact and AlmostExact graphs. Specifically, it makes sure that all IDs
in the same group should have the same extent. IDs with symbolic extents
are simply ignored. The check is executed only when
`EnableOption::IdModelExtraValidation` is turned on, which is true for
all the C++ tests.

The check indeed found a bug of the AlmostExact graph. Specifically,
when a broadcast is merged, we are mapping a non-broadcast input and its
output. This conceptually makes sense since merging with a broadcast
shouldn't change the concrete extent.

```
if (auto merge = dynamic_cast<Merge*>(expr)) {
    if (merge->inner()->isBroadcast() || merge->inner()->extent()->isOneInt()) {
      mapped_ids.push_back({merge->outer(), merge->out()});
    }
```

However, there's this behavior that I haven't considered. Specifically,
when a broadcast ID that has an extent larger than 1 is merged, the
merge output is also expanded by the factor of the broadcast extent.

```
  Split: bS2{1} by factor 4 -> bS3{1}, bS4{4}
  Merge: iS1{i0} and bS4{4} -> iS5{( i0 * 4 )}
```

In this case, the problem is the AlmostExact graph maps `iS1` and `iS5`,
but they shouldn't be mapped.

Maybe we should not create non-size-one broadcast IDs like `bS4` from
the beginning? Their extents don't make any sense, so maybe we should
enforce all broadcast IDs all have an extent of one?

TODO: Enable the validation option for the python tests