optimize pointwise, reduction, and normalizaiton schedulers for blackwell (#3754)
Summary of major changes:
**1. Pointwise scheduler:** Add unroll in inner or outer dim, unroll
factor is estimated from ops in the fusion & bytes in flight.
```
(1) Added function getUnrollFactor() to estimate unroll factor

// The unroll factor is calculated based on the following:
// (1) ensure enough bytes in flight to cover gmem access latency
// (2) ensure enough blocks for thread level parallelism (TLP)
// (3) don't unroll if split is not divisible and don't have enough TLP.


(2) Apply unroll to inner or outer dim
If the fusion is outer broadcast dominated, apply unroll to outer dim, otherwise, apply to inner dim.
Based on break point, when lhs byte multiple is smaller than rhs byte multiple, lhs has larger broadcast domains, the fusion is outer broadcast dominated.

(3) Revised bdimx to start from warp size to ensure divisible split of rhs elements in 2D scheduler. Performance profile is more smooth.
```
**2. Inner reduction scheduler:** 
```
(1) Add unroll on top of vectorization, unroll factor is derived from required bytes in flight.
(2) Prioritize block reduction over grid reduction, only do grid reduction when not all SMs are used.
(3) Revised to use a CTA size of \`128 or 256\`, and \`512\` is only used when the number of CTAs is small and we need more threads per CTA to saturate the SMs. This change leads to higher performance for fusions with computation expensive ops.
```


**3. Outer reduction scheduler**: Adjust unroll factor.
```
Previously, outer reduction unroll factor is derived using 8 buffer registers, after vectorization of 16 bytes, the unroll factor is 8 * 4 / 16 = 2. This is used when number of inputs <= 4. And decreased to 1 when the fusion has more than 4 inputs.
Intesad of using this dramastic change after 4 inputs, this PR changed to gradually reduce from 32 registers to 8 registers from 1 input to 4 inputs. For example, when there is 1 input, the corresponding register is 32, unroll factor is 32 x 4 / 16 = 8. Similarly, for 2 inputs, unroll factor is 32 * 4 / 16 / 2 = 4; 4 inputs, unroll factor is 2. This change only applies to fusions without expensive ops as test shows there are some regressions when the fusion has expensive ops, e.g. tanh, and no speedup is observed. The reason is because when we have expensive ops, computation cost of each thread is already high, we should prioritize thread level parallelism instead of instruction level parallelism.
```
**4. Inner persistent scheduler:**
```
(1) Added util function to return true if the device has a high bandiwidth to compute flops ratio. Based on test, the cut-off bandiwidth to compute flops ratio is set to 0.07. So it retures true for A100/B100/B200 and false for H100.

(2) When a device has high bandiwidth to compute flops ratio (A100 & B100/200), do the following optimizations:

\--(2.1) revise when should we disable project to inputs to avoid recomputation. Changed from 24KB to 96KB. Applies to softmax.

\--(2.2) reivse target occupany from 28 warps per sm to 16 warps per sm. Applies to softmax.

\--(2.3) disable magic 0. Applies to softmax

\--(2.4) revise sorting of candidates. Applies to all fusions, e.g. ln, rms norm, softmax.

The purposes of these optimizations are added in the code comments, generally they are added to reduce computation cost, increase persistent batch size (more bytes in flight)
```